Slowly working at making Swanki work with the Fluent Forever deck. This is going to involve writing my own Anki-Mustache-variant parser.

+++ Swanki/DeckConfig.swift
@@ -22,7 +22,7 @@ public struct LapseConfig: Codable {
   public let minInt: Int
   public let delays: [Int]
   public let leechAction: Int
-  public let mult: Int
+  public let mult: Double
 }
 
 public struct RevConfig: Codable {
